refactor: move types and utils into their own files
